Fix wrangler dev failures in site/docs by removing Node.js and WASM dependencies from Worker bundle
Split content.ts into runtime-safe types (content.ts) and build-time loader (content-loader.ts)
to eliminate node:fs/promises imports from the Worker bundle. Switch Shiki from default WASM engine
to JavaScript RegExp engine with fine-grained imports, matching site/ui's approach.
